+++ contract/src/error.rs
@@ -167,6 +167,8 @@ pub enum NFTCoreError {
     EmptyACLWhitelist = 162,
     InvalidACLPackageMode = 163,
     MissingACLPackageMode = 164,
+    InvalidPackageOperatorMode = 165,
+    MissingPackageOperatorMode = 166,
 }
 
 impl From<NFTCoreError> for ApiError {