+++ contract/src/main.rs
@@ -41,23 +41,24 @@ use constants::{
     ARG_HASH_KEY_NAME_1_0_0, ARG_HOLDER_MODE, ARG_IDENTIFIER_MODE, ARG_JSON_SCHEMA,
     ARG_METADATA_MUTABILITY, ARG_MINTING_MODE, ARG_NAMED_KEY_CONVENTION, ARG_NFT_KIND,
     ARG_NFT_METADATA_KIND, ARG_NFT_PACKAGE_KEY, ARG_OPERATOR, ARG_OPTIONAL_METADATA,
-    ARG_OWNERSHIP_MODE, ARG_OWNER_LOOKUP_MODE, ARG_RECEIPT_NAME, ARG_SOURCE_KEY, ARG_SPENDER,
-    ARG_TARGET_KEY, ARG_TOKEN_ID, ARG_TOKEN_META_DATA, ARG_TOKEN_OWNER, ARG_TOTAL_TOKEN_SUPPLY,
-    ARG_TRANSFER_FILTER_CONTRACT, ARG_WHITELIST_MODE, BURNT_TOKENS, BURN_MODE, COLLECTION_NAME,
-    COLLECTION_SYMBOL, ENTRY_POINT_APPROVE, ENTRY_POINT_BALANCE_OF, ENTRY_POINT_BURN,
-    ENTRY_POINT_GET_APPROVED, ENTRY_POINT_INIT, ENTRY_POINT_IS_APPROVED_FOR_ALL,
-    ENTRY_POINT_METADATA, ENTRY_POINT_MIGRATE, ENTRY_POINT_MINT, ENTRY_POINT_OWNER_OF,
-    ENTRY_POINT_REGISTER_OWNER, ENTRY_POINT_REVOKE, ENTRY_POINT_SET_APPROVALL_FOR_ALL,
-    ENTRY_POINT_SET_TOKEN_METADATA, ENTRY_POINT_SET_VARIABLES, ENTRY_POINT_TRANSFER,
-    ENTRY_POINT_UPDATED_RECEIPTS, EVENTS, EVENTS_MODE, HASH_BY_INDEX, HASH_KEY_NAME_1_0_0,
-    HOLDER_MODE, IDENTIFIER_MODE, INDEX_BY_HASH, INSTALLER, JSON_SCHEMA, MAX_TOTAL_TOKEN_SUPPLY,
-    METADATA_CEP78, METADATA_CUSTOM_VALIDATED, METADATA_MUTABILITY, METADATA_NFT721, METADATA_RAW,
-    MINTING_MODE, NFT_KIND, NFT_METADATA_KIND, NFT_METADATA_KINDS, NUMBER_OF_MINTED_TOKENS,
-    OPERATOR, OPERATORS, OWNED_TOKENS, OWNERSHIP_MODE, PAGE_LIMIT, PAGE_TABLE,
-    PREFIX_ACCESS_KEY_NAME, PREFIX_CEP78, PREFIX_CONTRACT_NAME, PREFIX_CONTRACT_VERSION,
-    PREFIX_HASH_KEY_NAME, PREFIX_PAGE_DICTIONARY, RECEIPT_NAME, REPORTING_MODE, RLO_MFLAG,
-    TOKEN_COUNT, TOKEN_ISSUERS, TOKEN_OWNERS, TOTAL_TOKEN_SUPPLY, TRANSFER_FILTER_CONTRACT,
-    TRANSFER_FILTER_CONTRACT_METHOD, UNMATCHED_HASH_COUNT, WHITELIST_MODE,
+    ARG_OWNERSHIP_MODE, ARG_OWNER_LOOKUP_MODE, ARG_PACKAGE_OPERATOR_MODE, ARG_RECEIPT_NAME,
+    ARG_SOURCE_KEY, ARG_SPENDER, ARG_TARGET_KEY, ARG_TOKEN_ID, ARG_TOKEN_META_DATA,
+    ARG_TOKEN_OWNER, ARG_TOTAL_TOKEN_SUPPLY, ARG_TRANSFER_FILTER_CONTRACT, ARG_WHITELIST_MODE,
+    BURNT_TOKENS, BURN_MODE, COLLECTION_NAME, COLLECTION_SYMBOL, ENTRY_POINT_APPROVE,
+    ENTRY_POINT_BALANCE_OF, ENTRY_POINT_BURN, ENTRY_POINT_GET_APPROVED, ENTRY_POINT_INIT,
+    ENTRY_POINT_IS_APPROVED_FOR_ALL, ENTRY_POINT_METADATA, ENTRY_POINT_MIGRATE, ENTRY_POINT_MINT,
+    ENTRY_POINT_OWNER_OF, ENTRY_POINT_REGISTER_OWNER, ENTRY_POINT_REVOKE,
+    ENTRY_POINT_SET_APPROVALL_FOR_ALL, ENTRY_POINT_SET_TOKEN_METADATA, ENTRY_POINT_SET_VARIABLES,
+    ENTRY_POINT_TRANSFER, ENTRY_POINT_UPDATED_RECEIPTS, EVENTS, EVENTS_MODE, HASH_BY_INDEX,
+    HASH_KEY_NAME_1_0_0, HOLDER_MODE, IDENTIFIER_MODE, INDEX_BY_HASH, INSTALLER, JSON_SCHEMA,
+    MAX_TOTAL_TOKEN_SUPPLY, METADATA_CEP78, METADATA_CUSTOM_VALIDATED, METADATA_MUTABILITY,
+    METADATA_NFT721, METADATA_RAW, MINTING_MODE, NFT_KIND, NFT_METADATA_KIND, NFT_METADATA_KINDS,
+    NUMBER_OF_MINTED_TOKENS, OPERATOR, OPERATORS, OWNED_TOKENS, OWNERSHIP_MODE,
+    PACKAGE_OPERATOR_MODE, PAGE_LIMIT, PAGE_TABLE, PREFIX_ACCESS_KEY_NAME, PREFIX_CEP78,
+    PREFIX_CONTRACT_NAME, PREFIX_CONTRACT_VERSION, PREFIX_HASH_KEY_NAME, PREFIX_PAGE_DICTIONARY,
+    RECEIPT_NAME, REPORTING_MODE, RLO_MFLAG, TOKEN_COUNT, TOKEN_ISSUERS, TOKEN_OWNERS,
+    TOTAL_TOKEN_SUPPLY, TRANSFER_FILTER_CONTRACT, TRANSFER_FILTER_CONTRACT_METHOD,
+    UNMATCHED_HASH_COUNT, WHITELIST_MODE,
 };
 use core::convert::{TryFrom, TryInto};
 use error::NFTCoreError;
@@ -201,6 +202,12 @@ pub extern "C" fn init() {
         NFTCoreError::InvalidACLPackageMode,
     )
     .unwrap_or_revert();
+
+    let package_operator_mode: bool = utils::get_optional_named_arg_with_user_errors::<bool>(
+        ARG_PACKAGE_OPERATOR_MODE,
+        NFTCoreError::InvalidPackageOperatorMode,
+    )
+    .unwrap_or_default();
 
     let receipt_name: String = utils::get_named_arg_with_user_errors(
         ARG_RECEIPT_NAME,
@@ -446,6 +453,10 @@ pub extern "C" fn init() {
     }
 
     runtime::put_key(ACL_PACKAGE_MODE, storage::new_uref(acl_package_mode).into());
+    runtime::put_key(
+        PACKAGE_OPERATOR_MODE,
+        storage::new_uref(package_operator_mode).into(),
+    );
 
     if vec![
         OwnerReverseLookupMode::Complete,
@@ -496,6 +507,7 @@ pub extern "C" fn set_variables() {
         );
         storage::write(allow_minting_uref, allow_minting);
     }
+
     if let Some(acl_package_mode) = utils::get_optional_named_arg_with_user_errors::<bool>(
         ARG_ACL_PACKAGE_MODE,
         NFTCoreError::MissingACLPackageMode,
@@ -506,6 +518,18 @@ pub extern "C" fn set_variables() {
             NFTCoreError::InvalidACLPackageMode,
         );
         storage::write(acl_package_mode_uref, acl_package_mode);
+    }
+
+    if let Some(package_operator_mode) = utils::get_optional_named_arg_with_user_errors::<bool>(
+        ARG_PACKAGE_OPERATOR_MODE,
+        NFTCoreError::MissingACLPackageMode,
+    ) {
+        let package_operator_mode_uref = utils::get_uref(
+            PACKAGE_OPERATOR_MODE,
+            NFTCoreError::MissingPackageOperatorMode,
+            NFTCoreError::InvalidPackageOperatorMode,
+        );
+        storage::write(package_operator_mode_uref, package_operator_mode);
     }
 
     let mut new_acl_whitelist = utils::get_optional_named_arg_with_user_errors::<Vec<Key>>(
@@ -917,10 +941,13 @@ pub extern "C" fn approve() {
         runtime::revert(NFTCoreError::InvalidOwnershipMode)
     }
 
-    let caller: Key = match utils::get_verified_caller().unwrap_or_revert() {
-        Caller::Session(account_hash) => account_hash.into(),
-        Caller::StoredCaller(contract_hash, _) => contract_hash.into(),
-    };
+    let (caller, contract_package): (Key, Option<Key>) =
+        match utils::get_verified_caller().unwrap_or_revert() {
+            Caller::Session(account_hash) => (account_hash.into(), None),
+            Caller::StoredCaller(contract_hash, contract_package_hash) => {
+                (contract_hash.into(), Some(contract_package_hash.into()))
+            }
+        };
 
     let identifier_mode: NFTIdentifierMode = utils::get_stored_value_with_user_errors::<u8>(
         IDENTIFIER_MODE,
@@ -966,8 +993,29 @@ pub extern "C" fn approve() {
         )
         .unwrap_or_default();
 
-    if !is_owner && !is_operator {
-        runtime::revert(NFTCoreError::InvalidAccountHash);
+    let is_package_operator = if !is_owner && !is_operator {
+        match (
+            utils::get_stored_value_with_user_errors::<bool>(
+                PACKAGE_OPERATOR_MODE,
+                NFTCoreError::MissingPackageOperatorMode,
+                NFTCoreError::InvalidPackageOperatorMode,
+            ),
+            contract_package,
+        ) {
+            (true, Some(contract_package)) => {
+                let owner_operator_item_key =
+                    utils::encode_key_and_value(&owner, &contract_package);
+                utils::get_dictionary_value_from_key::<bool>(OPERATORS, &owner_operator_item_key)
+                    .unwrap_or_default()
+            }
+            _ => false,
+        }
+    } else {
+        false
+    };
+
+    if !is_owner && !is_operator && !is_package_operator {
+        runtime::revert(NFTCoreError::InvalidTokenOwner);
     }
 
     // We assume a burnt token cannot be approved
@@ -1029,10 +1077,13 @@ pub extern "C" fn revoke() {
         runtime::revert(NFTCoreError::InvalidOwnershipMode)
     }
 
-    let caller: Key = match utils::get_verified_caller().unwrap_or_revert() {
-        Caller::Session(account_hash) => account_hash.into(),
-        Caller::StoredCaller(contract_hash, _) => contract_hash.into(),
-    };
+    let (caller, contract_package): (Key, Option<Key>) =
+        match utils::get_verified_caller().unwrap_or_revert() {
+            Caller::Session(account_hash) => (account_hash.into(), None),
+            Caller::StoredCaller(contract_hash, contract_package_hash) => {
+                (contract_hash.into(), Some(contract_package_hash.into()))
+            }
+        };
 
     let identifier_mode: NFTIdentifierMode = utils::get_stored_value_with_user_errors::<u8>(
         IDENTIFIER_MODE,
@@ -1078,8 +1129,29 @@ pub extern "C" fn revoke() {
         )
         .unwrap_or_default();
 
-    if !is_owner && !is_operator {
-        runtime::revert(NFTCoreError::InvalidAccountHash);
+    let is_package_operator = if !is_owner && !is_operator {
+        match (
+            utils::get_stored_value_with_user_errors::<bool>(
+                PACKAGE_OPERATOR_MODE,
+                NFTCoreError::MissingPackageOperatorMode,
+                NFTCoreError::InvalidPackageOperatorMode,
+            ),
+            contract_package,
+        ) {
+            (true, Some(contract_package)) => {
+                let owner_operator_item_key =
+                    utils::encode_key_and_value(&owner, &contract_package);
+                utils::get_dictionary_value_from_key::<bool>(OPERATORS, &owner_operator_item_key)
+                    .unwrap_or_default()
+            }
+            _ => false,
+        }
+    } else {
+        false
+    };
+
+    if !is_owner && !is_operator && !is_package_operator {
+        runtime::revert(NFTCoreError::InvalidTokenOwner);
     }
 
     // We assume a burnt token cannot be revoked
@@ -1260,29 +1332,58 @@ pub extern "C" fn transfer() {
         runtime::revert(NFTCoreError::InvalidAccount);
     }
 
-    let caller: Key = match utils::get_verified_caller().unwrap_or_revert() {
-        Caller::Session(account_hash) => account_hash.into(),
-        Caller::StoredCaller(contract_hash, _) => contract_hash.into(),
-    };
+    let (caller, contract_package): (Key, Option<Key>) =
+        match utils::get_verified_caller().unwrap_or_revert() {
+            Caller::Session(account_hash) => (account_hash.into(), None),
+            Caller::StoredCaller(contract_hash, contract_package_hash) => {
+                (contract_hash.into(), Some(contract_package_hash.into()))
+            }
+        };
 
     // Check if caller is owner
     let is_owner = owner == caller;
 
     // Check if caller is approved to execute transfer
-    let is_approved = match utils::get_dictionary_value_from_key::<Option<Key>>(
-        APPROVED,
-        &token_identifier.get_dictionary_item_key(),
-    ) {
-        Some(Some(maybe_approved)) => caller == maybe_approved,
-        Some(None) | None => false,
-    };
+    let is_approved = !is_owner
+        && match utils::get_dictionary_value_from_key::<Option<Key>>(
+            APPROVED,
+            &token_identifier.get_dictionary_item_key(),
+        ) {
+            Some(Some(maybe_approved)) => caller == maybe_approved,
+            Some(None) | None => false,
+        };
 
     // Check if caller is operator to execute transfer
-    let owner_operator_item_key = utils::encode_key_and_value(&source_owner_key, &caller);
+    let is_operator = if !is_owner && !is_approved {
+        let owner_operator_item_key = utils::encode_key_and_value(&source_owner_key, &caller);
+        utils::get_dictionary_value_from_key::<bool>(OPERATORS, &owner_operator_item_key)
+            .unwrap_or_default()
+    } else {
+        false
+    };
 
-    let is_operator = !is_approved
-        && utils::get_dictionary_value_from_key::<bool>(OPERATORS, &owner_operator_item_key)
-            .unwrap_or_default();
+    // With operator package mode check if caller's package is operator to let contract execute
+    // transfer
+    let is_package_operator = if !is_owner && !is_approved && !is_operator {
+        match (
+            utils::get_stored_value_with_user_errors::<bool>(
+                PACKAGE_OPERATOR_MODE,
+                NFTCoreError::MissingPackageOperatorMode,
+                NFTCoreError::InvalidPackageOperatorMode,
+            ),
+            contract_package,
+        ) {
+            (true, Some(contract_package)) => {
+                let owner_operator_item_key =
+                    utils::encode_key_and_value(&source_owner_key, &contract_package);
+                utils::get_dictionary_value_from_key::<bool>(OPERATORS, &owner_operator_item_key)
+                    .unwrap_or_default()
+            }
+            _ => false,
+        }
+    } else {
+        false
+    };
 
     if let Some(filter_contract) = utils::get_transfer_filter_contract() {
         let mut args = RuntimeArgs::new();
@@ -1306,7 +1407,7 @@ pub extern "C" fn transfer() {
     }
 
     // Revert if caller is not owner nor approved nor an operator.
-    if !is_owner && !is_approved && !is_operator {
+    if !is_owner && !is_approved && !is_operator && !is_package_operator {
         runtime::revert(NFTCoreError::InvalidTokenOwner);
     }
 
@@ -1860,9 +1961,23 @@ pub extern "C" fn migrate() {
     )
     .unwrap_or_default();
 
-    // Don't overwrite stored value if acl package mode is not a param or false
+    // Don't overwrite stored value on migration if acl package mode is missing param or false
     if acl_package_mode {
         runtime::put_key(ACL_PACKAGE_MODE, storage::new_uref(acl_package_mode).into());
+    }
+
+    let package_operator_mode: bool = utils::get_optional_named_arg_with_user_errors::<bool>(
+        ARG_PACKAGE_OPERATOR_MODE,
+        NFTCoreError::InvalidPackageOperatorMode,
+    )
+    .unwrap_or_default();
+
+    // Don't overwrite stored value on migration if package operator mode is missing param or false
+    if package_operator_mode {
+        runtime::put_key(
+            PACKAGE_OPERATOR_MODE,
+            storage::new_uref(package_operator_mode).into(),
+        );
     }
 
     // Duplicate old dict OPERATOR named key to new dict APPROVED named key
@@ -2015,6 +2130,7 @@ fn generate_entry_points() -> EntryPoints {
             Parameter::new(ARG_WHITELIST_MODE, CLType::U8),
             Parameter::new(ARG_ACL_WHITELIST, CLType::List(Box::new(CLType::Key))),
             Parameter::new(ARG_ACL_PACKAGE_MODE, CLType::Bool),
+            Parameter::new(ARG_PACKAGE_OPERATOR_MODE, CLType::Bool),
             Parameter::new(ARG_JSON_SCHEMA, CLType::String),
             Parameter::new(ARG_RECEIPT_NAME, CLType::String),
             Parameter::new(ARG_IDENTIFIER_MODE, CLType::U8),
@@ -2053,6 +2169,7 @@ fn generate_entry_points() -> EntryPoints {
             ),
             Parameter::new(ARG_ACL_WHITELIST, CLType::List(Box::new(CLType::Key))),
             Parameter::new(ARG_ACL_PACKAGE_MODE, CLType::Bool),
+            Parameter::new(ARG_PACKAGE_OPERATOR_MODE, CLType::Bool),
         ],
         CLType::Unit,
         EntryPointAccess::Public,
@@ -2398,6 +2515,12 @@ fn install_contract() {
     )
     .unwrap_or_default();
 
+    let package_operator_mode: bool = utils::get_optional_named_arg_with_user_errors::<bool>(
+        ARG_PACKAGE_OPERATOR_MODE,
+        NFTCoreError::InvalidPackageOperatorMode,
+    )
+    .unwrap_or_default();
+
     // Represents the schema for the metadata for a given NFT contract instance.
     // Refer to the `NFTMetadataKind` enum in src/utils for details.
     // This value cannot be changed after installation.
@@ -2552,6 +2675,7 @@ fn install_contract() {
         ARG_NFT_PACKAGE_KEY => nft_contract_package_hash.to_formatted_string(),
         ARG_EVENTS_MODE => events_mode,
         ARG_ACL_PACKAGE_MODE => acl_package_mode,
+        ARG_PACKAGE_OPERATOR_MODE => package_operator_mode,
         ARG_TRANSFER_FILTER_CONTRACT =>
         transfer_filter_contract_contract_key,
     };
@@ -2614,10 +2738,17 @@ fn migrate_contract(access_key_name: String, package_key_name: String) {
     )
     .unwrap_or_default();
 
+    let package_operator_mode: bool = utils::get_optional_named_arg_with_user_errors::<bool>(
+        ARG_PACKAGE_OPERATOR_MODE,
+        NFTCoreError::InvalidPackageOperatorMode,
+    )
+    .unwrap_or_default();
+
     let mut runtime_args = runtime_args! {
         ARG_NFT_PACKAGE_KEY => nft_contract_package_hash,
         ARG_EVENTS_MODE => events_mode,
         ARG_ACL_PACKAGE_MODE => acl_package_mode,
+        ARG_PACKAGE_OPERATOR_MODE => package_operator_mode,
     };
 
     if let Some(new_token_supply) = utils::get_optional_named_arg_with_user_errors::<u64>(