+++ tests/src/upgrade.rs
@@ -8,10 +8,10 @@ use contract::{
     constants::{
         ACCESS_KEY_NAME_1_0_0, ACL_PACKAGE_MODE, ARG_ACCESS_KEY_NAME_1_0_0, ARG_ACL_PACKAGE_MODE,
         ARG_COLLECTION_NAME, ARG_EVENTS_MODE, ARG_HASH_KEY_NAME_1_0_0, ARG_NAMED_KEY_CONVENTION,
-        ARG_SOURCE_KEY, ARG_TARGET_KEY, ARG_TOKEN_HASH, ARG_TOKEN_META_DATA, ARG_TOKEN_OWNER,
-        ARG_TOTAL_TOKEN_SUPPLY, ENTRY_POINT_MINT, ENTRY_POINT_REGISTER_OWNER,
-        NUMBER_OF_MINTED_TOKENS, PAGE_LIMIT, PREFIX_ACCESS_KEY_NAME, PREFIX_HASH_KEY_NAME,
-        RECEIPT_NAME, UNMATCHED_HASH_COUNT,
+        ARG_PACKAGE_OPERATOR_MODE, ARG_SOURCE_KEY, ARG_TARGET_KEY, ARG_TOKEN_HASH,
+        ARG_TOKEN_META_DATA, ARG_TOKEN_OWNER, ARG_TOTAL_TOKEN_SUPPLY, ENTRY_POINT_MINT,
+        ENTRY_POINT_REGISTER_OWNER, NUMBER_OF_MINTED_TOKENS, PACKAGE_OPERATOR_MODE, PAGE_LIMIT,
+        PREFIX_ACCESS_KEY_NAME, PREFIX_HASH_KEY_NAME, RECEIPT_NAME, UNMATCHED_HASH_COUNT,
     },
     events::events_ces::Migration,
     modalities::EventsMode,
@@ -869,6 +869,69 @@ fn should_safely_upgrade_with_acl_package_mode() {
     );
 
     assert!(is_acl_packge_mode);
+
+    // Expect Migration event.
+    let expected_event = Migration::new();
+    let actual_event: Migration = support::get_event(&builder, &nft_contract_key, 0);
+    assert_eq!(actual_event, expected_event, "Expected Migration event.");
+}
+
+#[test]
+fn should_safely_upgrade_with_package_operator_mode() {
+    let mut builder = InMemoryWasmTestBuilder::default();
+    builder
+        .run_genesis(&PRODUCTION_RUN_GENESIS_REQUEST)
+        .commit();
+
+    let install_request = InstallerRequestBuilder::new(*DEFAULT_ACCOUNT_ADDR, CONTRACT_1_0_0_WASM)
+        .with_collection_name(NFT_TEST_COLLECTION.to_string())
+        .with_collection_symbol(NFT_TEST_SYMBOL.to_string())
+        .with_total_token_supply(1000u64)
+        .with_ownership_mode(OwnershipMode::Minter)
+        .with_identifier_mode(NFTIdentifierMode::Ordinal)
+        .with_nft_metadata_kind(NFTMetadataKind::Raw)
+        .build();
+
+    builder.exec(install_request).expect_success().commit();
+
+    let nft_contract_hash_1_0_0 = support::get_nft_contract_hash_1_0_0(&builder);
+    let nft_contract_key_1_0_0: Key = nft_contract_hash_1_0_0.into();
+
+    let is_package_operator_mode = builder
+        .query(None, nft_contract_key_1_0_0, &[])
+        .expect("must have nft contract")
+        .as_contract()
+        .expect("must convert contract")
+        .named_keys()
+        .contains_key(PACKAGE_OPERATOR_MODE);
+
+    assert!(!is_package_operator_mode);
+
+    let upgrade_request = ExecuteRequestBuilder::standard(
+        *DEFAULT_ACCOUNT_ADDR,
+        NFT_CONTRACT_WASM,
+        runtime_args! {
+            ARG_NFT_CONTRACT_HASH => nft_contract_hash_1_0_0,
+            ARG_COLLECTION_NAME => NFT_TEST_COLLECTION.to_string(),
+            ARG_NAMED_KEY_CONVENTION => NamedKeyConventionMode::V1_0Standard as u8,
+            ARG_EVENTS_MODE => EventsMode::CES as u8,
+            ARG_PACKAGE_OPERATOR_MODE => true,
+        },
+    )
+    .build();
+
+    builder.exec(upgrade_request).expect_success().commit();
+
+    let nft_contract_hash = support::get_nft_contract_hash(&builder);
+    let nft_contract_key: Key = nft_contract_hash.into();
+
+    let is_package_operator_mode: bool = support::query_stored_value(
+        &builder,
+        nft_contract_key,
+        vec![PACKAGE_OPERATOR_MODE.to_string()],
+    );
+
+    assert!(is_package_operator_mode);
 
     // Expect Migration event.
     let expected_event = Migration::new();